fixed PR comments, added a push range function to concurrent queue, and fixed bug in chromatic scheduling

+++ src/libgraphio/libgraphio.h
@@ -4,7 +4,7 @@
 #include <string>
 #include <cinttypes>
 
-typedef int64_t vid_t;  // vertex id type
+typedef int32_t vid_t;  // vertex id type
 
 class EdgeListBuilder {
  public:
